requirements.txt
fastapi>=0.100.0
tortoise-orm[asyncpg]>=0.20.0
uvicorn[standard]>=0.23.0
pydantic[email]>=2.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
pytest>=7.4.0
httpx>=0.24.0
aerich>=0.7.2
slowapi>=0.1.9
python-decouple>=3.8
aiosqlite>=0.21.0
jinja2>=3.1.0
python-multipart>=0.0.6